Create csvtoxml.py
silly csv to xml converter in python, uses a token list and tag list to cycle through lines in a csv file and convert it to xml... supports attributes and values... hard to use, i should write a gui some day. In any case, I used it to do what I needed to do, on more than one occasion.